# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bookstore API",
    "description": "API for managing bookstore, users, and books.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8890",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "This endpoint allows new users to register in the system.",
        "operationId": "registerUser",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "john_doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User registered successfully!"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "john_doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "john@example.com"
                        },
                        "created_date": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-11-05T12:00:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, missing fields or validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login a user",
        "description": "This endpoint allows a user to log in and receive a JWT token.",
        "operationId": "loginUser",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, JWT token returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "john_doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "john@example.com"
                        },
                        "created_date": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-11-05T12:00:00Z"
                        }
                      }
                    },
                    "token": {
                      "type": "string",
                      "example": "YOUR_JWT_TOKEN"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "summary": "Get all books",
        "description": "This endpoint returns a list of all books with pagination and filtering.",
        "operationId": "getBooks",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (default is 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of books per page (default is 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "title",
            "in": "query",
            "description": "Filter by book title",
            "required": false,
            "schema": {
              "type": "string",
              "example": "gatsby"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "Filter by book author",
            "required": false,
            "schema": {
              "type": "string",
              "example": "fitzgerald"
            }
          },
          {
            "name": "isbn",
            "in": "query",
            "description": "Filter by book ISBN",
            "required": false,
            "schema": {
              "type": "string",
              "example": "9780743273565"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of books with pagination and filters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalBooks": {
                      "type": "integer",
                      "example": 50
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 5
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "books": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "title": {
                            "type": "string",
                            "example": "The Great Gatsby"
                          },
                          "author": {
                            "type": "string",
                            "example": "F. Scott Fitzgerald"
                          },
                          "isbn": {
                            "type": "string",
                            "example": "9780743273565"
                          },
                          "published_date": {
                            "type": "string",
                            "format": "date",
                            "example": "1925-04-10"
                          },
                          "pages": {
                            "type": "integer",
                            "example": 180
                          },
                          "language": {
                            "type": "string",
                            "example": "English"
                          },
                          "publisher": {
                            "type": "string",
                            "example": "Scribner"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch books"
          }
        }
      },
      "post": {
        "summary": "Create a new book",
        "description": "This endpoint allows creating a new book in the bookstore.",
        "operationId": "createBook",
        "tags": ["Books"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "The Great Gatsby"
                  },
                  "author": {
                    "type": "string",
                    "example": "F. Scott Fitzgerald"
                  },
                  "isbn": {
                    "type": "string",
                    "example": "9780743273565"
                  },
                  "published_date": {
                    "type": "string",
                    "format": "date",
                    "example": "1925-04-10"
                  },
                  "pages": {
                    "type": "integer",
                    "example": 180
                  },
                  "language": {
                    "type": "string",
                    "example": "English"
                  },
                  "publisher": {
                    "type": "string",
                    "example": "Scribner"
                  }
                },
                "required": ["title", "author", "isbn"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "The Great Gatsby"
                    },
                    "author": {
                      "type": "string",
                      "example": "F. Scott Fitzgerald"
                    },
                    "isbn": {
                      "type": "string",
                      "example": "9780743273565"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, missing or invalid fields"
          },
          "401": {
            "description": "Unauthorized, JWT token required"
          },
          "500": {
            "description": "Failed to create book"
          }
        }
      }
    },
    
    "/api/books/{id}": {
      "get": {
        "summary": "Get a book",
        "description": "This endpoint allows get a book.",
        "operationId": "getBook",
        "tags": ["Books"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to update",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get Book successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "The Great Gatsby (Updated)"
                    },
                    "author": {
                      "type": "string",
                      "example": "F. Scott Fitzgerald"
                    },
                    "isbn": {
                      "type": "string",
                      "example": "9780743273565"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid fields"
          },
          "401": {
            "description": "Unauthorized, JWT token required"
          }
        }
      },
      "put": {
        "summary": "Update a book",
        "description": "This endpoint allows updating an existing book.",
        "operationId": "updateBook",
        "tags": ["Books"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to update",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "The Great Gatsby (Updated)"
                  },
                  "author": {
                    "type": "string",
                    "example": "F. Scott Fitzgerald"
                  },
                  "isbn": {
                    "type": "string",
                    "example": "9780743273565"
                  },
                  "published_date": {
                    "type": "string",
                    "format": "date",
                    "example": "1925-04-10"
                  },
                  "pages": {
                    "type": "integer",
                    "example": 180
                  },
                  "language": {
                    "type": "string",
                    "example": "English"
                  },
                  "publisher": {
                    "type": "string",
                    "example": "Scribner"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "The Great Gatsby (Updated)"
                    },
                    "author": {
                      "type": "string",
                      "example": "F. Scott Fitzgerald"
                    },
                    "isbn": {
                      "type": "string",
                      "example": "9780743273565"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid fields"
          },
          "401": {
            "description": "Unauthorized, JWT token required"
          },
          "500": {
            "description": "Failed to update book"
          }
        }
      },
      "delete": {
        "summary": "Delete a book",
        "description": "This endpoint allows deleting a book from the bookstore.",
        "operationId": "deleteBook",
        "tags": ["Books"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the book to delete",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book successfully deleted"
          },
          "401": {
            "description": "Unauthorized, JWT token required"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Failed to delete book"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}